Alpha Release v1.0
This Alpha Release of the worlds first Private Virtual Currency (Bitcoin) Securities Exchange was a module within a web service entitled Wave. The two are to be separated and so the first milestone goal of this project is to strip away everything which doesn't pertain to the virtual currency securities exchange. The overall goal of this newly open-sourced securities exchange is to give webmasters a portal for their websites whereby they are able to split ownership of their web service in shares, then sell off equity in exchange for bitcoins. The investors will need to be able to access the portal to trade equity and bitcoins at will (similar to Kraken)